requirements.txt
Django==5.2
Pillow==11.2.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-dotenv==1.0.1
django-jazzmin==3.0.1
django-filter==24.1
django-widget-tweaks==1.5.0
django-allauth==0.61.1
xhtml2pdf==0.2.13
django-debug-toolbar==4.3.0
whitenoise==6.6.0
django-ckeditor==6.7.0
django-tables2==2.7.0
djangorestframework==3.14.0
